Move `HTTPClient` to its own file
